requirements.txt
requests
beautifulsoup4
trycourier
google-generativeai
rich
